+++ src/package/constants/process.ts
@@ -1,7 +1,7 @@
 /**
  * Exit status of e2ed process.
  */
-export const enum ExitStatus {
+export enum ExitStatus {
   Passed,
   Failed,
   NoRetries,