+++ src/package/environment.ts
@@ -3,7 +3,7 @@ const RUN_ENVIRONMENT_VARIABLE_NAME = '__INTERNAL_E2ED_RUN_ENVIRONMENT';
 /**
  * Run environment enum (run in docker or local run).
  */
-export const enum RunEnvironment {
+export enum RunEnvironment {
   Docker = 'docker',
   Local = 'local',
 }